Release a pretrained Inception v3 model checkpoint.
The model was trained on the publicly released annotations and
is expected to serve as a baseline. It's possible to boost performance
by reducing the noise in the annotatoins, improving training procedure
and calibrating the output.

This release also makes it possible to accept external contributions
to the repository under Apache 2.0 license. See CONTRIBUTING.md for
more details.

Fixes #3.